nop IoSystem for benchmarking
